+++ .gitignore
@@ -1,3 +1,6 @@
+# jetBrains
+.idea
+
 # Logs
 logs
 *.log
