+++ vue-frontend/src/components/Quiz Creation/MCQuestion.vue
@@ -2,9 +2,10 @@
   const props = defineProps({
     question: String,
     answers: Array
-    
   })
 </script>
 <template>
-
+  <div class="rounded-md border-2 mx-4 my-4">
+    something something
+  </div>
 </template>
