--- vue-frontend/src/views/CreateQuizzes/CreateMCQ.vue
@@ -1,6 +1,0 @@
-<script setup>
-</script>
-
-<template>
-  This is MCQ
-</template>
