+++ vue-frontend/src/views/views.js
@@ -3,7 +3,6 @@ export { default as AboutView } from './AboutView.vue'
 export { default as HomeView } from './HomeView.vue'
 export { default as LoginView } from './LoginView.vue'
 export { default as SignUpView } from './SignUpView.vue'
-export { default as CreateQuizView } from './CreateQuizzes/CreateQuizView.vue'
-export { default as CreateMCQ } from './CreateQuizzes/CreateMCQ.vue'
+export { default as CreateQuizView } from './CreateQuizView.vue'
 export { default as ForgotPasswordView } from './ForgotPasswordView.vue'
 export { default as StudentHomeView } from './StudentHomeView.vue'
